remove merge conflicts

+++ src/atom_vec_bdipole.cpp
@@ -871,9 +871,9 @@ int AtomVecBDipole::write_data_hybrid(FILE *fp, double *buf)
    return # of bytes of allocated memory
 ------------------------------------------------------------------------- */
 
-bigint AtomVecBDipole::memory_usage()
+double AtomVecBDipole::memory_usage()
 {
-  bigint bytes = 0;
+  double bytes = 0;
 
   if (atom->memcheck("tag")) bytes += memory->usage(tag,nmax);
   if (atom->memcheck("type")) bytes += memory->usage(type,nmax);
