+++ src/atom_vec_bdipole.h
@@ -56,7 +56,7 @@ class AtomVecBDipole : public AtomVec {
   int pack_data_hybrid(int, double *);
   void write_data(FILE *, int, double **);
   int write_data_hybrid(FILE *, double *);
-  bigint memory_usage();
+  double memory_usage();
 
  private:
   tagint *tag;
